+++ Sources/RadioKit/Extensions/Image+MediaItem.swift
@@ -16,16 +16,28 @@ import MediaPlayer
 #if canImport(UIKit)
   extension UIImage {
     /// A Now Playing artwork wrapper that renders this image at any requested size.
-    public var artwork: MPMediaItemArtwork {
+    public var mediaItemArtwork: MPMediaItemArtwork {
       MPMediaItemArtwork(boundsSize: size) { _ in self }
+    }
+
+    /// The previous name of `mediaItemArtwork`.
+    @available(*, deprecated, renamed: "mediaItemArtwork")
+    public var artwork: MPMediaItemArtwork {
+      mediaItemArtwork
     }
   }
 
 #elseif canImport(AppKit)
   extension NSImage {
     /// A Now Playing artwork wrapper that renders this image at any requested size.
-    public var artwork: MPMediaItemArtwork {
+    public var mediaItemArtwork: MPMediaItemArtwork {
       MPMediaItemArtwork(boundsSize: size) { _ in self }
+    }
+
+    /// The previous name of `mediaItemArtwork`.
+    @available(*, deprecated, renamed: "mediaItemArtwork")
+    public var artwork: MPMediaItemArtwork {
+      mediaItemArtwork
     }
   }
 #endif